built-store-paths: finish pbh code rearrangement
